feat(motion-matching/option2): relocate per-step surrogate to option2_nn_surrogate/perstep (#4044)
Closes #4044. Moves train_dynamics_surrogate.py, optimize_torque_sequence_for_club.py
and extract_dynamics_dataset.py from MachineLearning/ to
src/shared/python/motion_matching/surrogate/perstep/. Backwards-compat shims
preserve the old import paths with deprecation warnings. Updates APPROACH.md
with a per-step vs trajectory-level tradeoff table. Adds relocation tests
plus a slow-marked held-out RMSE comparison test.
